+++ .vscode-test.js
@@ -1,0 +1,6 @@
+const { defineConfig } = require('@vscode/test-cli');
+
+module.exports = defineConfig([{ 
+  files: '**/suite/*.test.js',
+  version: 'insiders',
+}]);